+++ aldrin/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `ChannelCookie`s.
 - Add `client()` getters to `Sender` and `Receiver`.
 - Add `introspection` Cargo feature to enable introspection support.
+- Add `RunError::Serialize` variant.
 
 ## [0.6.0] - 2024-06-07